Update boundary_elements.py
(1) extract the integer values only from the flowpath_id naming convention in the NextGen framework for coastal model BMI use, assign the variable as "POI" to include in the source_sink.in file.
(2) After consulting with the GIS team, the "true" unique identifiers within the inflow_boundaries.csv file is the "hl_link" column constructed. Revised the loop from "flowpath_id" to "hl_link" to reflect the true unique identifiers to link with intersecting SCHISM elements
(3) Assing the POI field with the corresponding element id to the source_sink_BMI.in file to properly construct the SCHISM BMI dependencies.